Move `common/auth.go` to `common/integrations/auth/`
Split it to several files
