+++ backend/chat/handlers/handlers.go
@@ -4,7 +4,7 @@ package handlers
 import (
 	"backend/chat/database"
 	"backend/chat/ws"
-	"backend/common/external"
+	"backend/common/integrations/auth"
 	"database/sql"
 	"encoding/json"
 	"log"
@@ -47,7 +47,7 @@ func (h *Handlers) PrimaryHandler() ws.MessageHandler {
 		}
 
 		// TODO: maybe remove, as we have a revoked-sessions endpoint
-		userId := external.CheckSession(client.Token)
+		userId := auth.CheckSession(client.Token)
 		if userId != client.UserId {
 			log.Println("ERROR: invalid token")
 			_ = client.Conn.Close()
